Rebased event_functionality onto skill_functionality, refactored code to reflect the reorganized models packages.
