+++ src/main/java/com/skills/skills/controllers/AuthenticationController.java
@@ -3,12 +3,12 @@ package com.skills.skills.controllers;
 import com.skills.skills.data.SkillsCategoryRepository;
 import com.skills.skills.data.SkillsRepository;
 import com.skills.skills.data.UserRepository;
-import com.skills.skills.models.Skill;
-import com.skills.skills.models.SkillsCategory;
-import com.skills.skills.models.User;
-import com.skills.skills.models.UserProfile;
+import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import com.skills.skills.models.dto.LoginFormDTO;
 import com.skills.skills.models.dto.RegisterFormDTO;
+import com.skills.skills.models.user.User;
+import com.skills.skills.models.user.UserProfile;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
