+++ src/main/java/com/skills/skills/data/SkillsCategoryRepository.java
@@ -1,6 +1,6 @@
 package com.skills.skills.data;
 
-import com.skills.skills.models.SkillsCategory;
+import com.skills.skills.models.skill.SkillsCategory;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
